Add attribute to be able to control the bitbucket service
If the new attribute will stay set to 'true', no changes will be made to the cookbook or the server installation.
If it will be changed to 'false' -
1. The bitbucket service won't be enabled on boot.
2. All restart notifications from config files won't really restart the service.
3. Future start\restart notification won't change the service's state.

All those will result in a installed bitbucket server, but with stopped service.
And basically it will now enable installing a cold standby server as described here:
https://confluence.atlassian.com/bitbucketserver/high-availability-for-bitbucket-server-776640137.html